requirements.txt
streamlit==1.45.0
pandas==2.2.3
numpy==2.2.4
plotly==6.1.2
scikit-learn==1.6.1
joblib==1.4.2
xgboost==3.0.2
torch>=2.7.0
torchvision>=0.22.0
Pillow>=11.1.0
